Move .macro → .macro.calibrate
- Adjust imports.
- Adjust tests.
