+++ message_ix/core.py
@@ -793,7 +793,7 @@ class Scenario(ixmp.Scenario):
         --------
         :ref:`macro-input-data`
         """
-        from .macro import EXPERIMENTAL, add_model_data, calibrate
+        from .macro.calibrate import EXPERIMENTAL, add_model_data, calibrate
         from .models import MACRO
 
         # Display a warning
